Moved replacers to own file & Added disable functionality
Moved replacers to own file.
Added disable functionality.
Added functionality to add or remove replacers via tsconfig
